get rid of more `Ixxx` classes
